==> data-science-algo/tasks/3_model_evaluation.ipynb <==
# Оценка модели
Строим модель энергопотребления здания по среднему значению. Оцениваем эффективность модели через метрику
$$\LARGE RMSLE = \sqrt{\frac{\sum_{i=1}^{n}(log(p_i+1)-log(a_i+1))^2}{n}},$$
где:
- $n - число\; наблюдений;$
- $log\;- натуральный\; логорифм;$
- $p_i\;- вычисленное\; значение\; метрики;$
- $a_i\;- заданное\; значение\; метрики.$

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

## Загрузка данных. Выкунуть пустые дни по "meter_reading" и выделить час в отдельную колонку.
**meter_reading** - потребление энергии в кВтч (или эквивалент)

energy = pd.read_csv("../data/train.csv")
energy = energy[energy["meter_reading"]>0]
energy["timestamp"] = pd.to_datetime(energy["timestamp"])
energy["hour"] = energy["timestamp"].dt.hour
print(energy.head())

# Разделить данные на обучение и проверку
**20%** - на проверку
**80%** - на обучение

energy_train, energy_test = train_test_split(energy, test_size=0.2)
print(energy_train.count())

## Среднее и медианное значение потребления энергии по часам

energy_train_hours = energy_train.groupby("hour")
energy_meter_reading = energy_train_hours["meter_reading"]
energy_train_avg = pd.DataFrame(
    {
        "Среднее": energy_meter_reading.mean(),
        "Медиана": energy_meter_reading.median(),
    }
)
print(energy_train_avg)

## Проверка модели

def model(x):
    meter_reading_log = np.log(x.meter_reading + 1)
    meter_reading_mean = np.log(energy_train_avg["Среднее"][x.hour] + 1)
    meter_reading_median = np.log(energy_train_avg["Медиана"][x.hour] + 1)
    
    x["meter_reading_mean_q"] = (meter_reading_log - meter_reading_mean)**2
    x["meter_reading_median_q"] = (meter_reading_log - meter_reading_median)**2
    x["meter_reading_zero_q"] = meter_reading_log**2

    return x

energy_test = energy_test.apply(model, axis=1, result_type="expand")
print(energy_test.head())

len_ = len(energy_test)
energy_test_median_rmsle = np.sqrt(energy_test["meter_reading_median_q"].sum() / len_)
energy_test_mean_rmsle = np.sqrt(energy_test["meter_reading_mean_q"].sum() / len_)
energy_test_zero_rmsle = np.sqrt(energy_test["meter_reading_zero_q"].sum() / len_)

print(f"Качество медианы: {energy_test_median_rmsle}")
print(f"Качество среднего: {energy_test_mean_rmsle}")
print(f"Качество нуля: {energy_test_zero_rmsle}")
