==> data-science-algo/tasks/4_linear_regression.ipynb <==
# Линейная регрессия
Модель линейной регрессии энергопотребления здания, используя температуру воздуха и влажность.

%matplotlib inline

import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression

## Загрузка, объединение, фильтрация данных

buildings= pd.read_csv("../data/buildings.csv")
weather = pd.read_csv("../data/weather.csv")
energy = pd.read_csv("../data/train.csv")

print(energy.info())

energy = pd.merge(left=energy, right=buildings, how="left", left_on="building_id", right_on="building_id")
energy.set_index(["timestamp", "site_id"], inplace=True)

weather.set_index(["timestamp", "site_id"], inplace=True)
energy = pd.merge(left=energy, right=weather, how="left", left_index=True, right_index=True)
energy.reset_index(inplace=True)

energy = energy[energy["meter_reading"]>0]
energy["timestamp"] = pd.to_datetime(energy["timestamp"])
energy["hour"] = energy["timestamp"].dt.hour

print(energy.head())


## Разделение данных на обучение и проверку

energy_train, energy_test = train_test_split(energy, test_size=0.2)
print(energy_train.head())

## Модель линейной регрессии и среднее

$$\normalsize meter\_reading = A*air\_temperature+B*air\_temperature+C$$

где:
- $meter\_reading - потребление\; энергии\; в\; кВтч\; (или\; эквивалент);$
- $air\_temperature - температура\; воздуха, С;$
- $air\_temperature - точка\; росы\; (влажность), С;$

Дополнительно вычислим среднее по часам, что бы сравнить линейную регрессию с более простой моделью

energy_train_avg = energy_train.groupby("hour")["meter_reading"].mean()

energy_train_lr = pd.DataFrame(
    energy_train,
    columns=["meter_reading", "air_temperature", "dew_temperature"]
)

y = energy_train_lr["meter_reading"]
x = energy_train_lr.drop(labels=["meter_reading"], axis=1)
model = LinearRegression().fit(x,y)

print(model.coef_, model.intercept_)

## Оценка модели

def calculate_model(x):
    meter_reading_log = np.log(x.meter_reading + 1)
    meter_reading_mean = np.log(energy_train_avg[x.hour] + 1)
    meter_reading_lr = np.log(
        1 + x.air_temperature * model.coef_[0] +
        x.dew_temperature * model.coef_[1] +
        model.intercept_
    )
    x["meter_reading_lr_q"] = (meter_reading_log - meter_reading_lr)**2
    x["meter_reading_mean_q"] = (meter_reading_log - meter_reading_mean)**2

    return x

len_ = len(energy_test)
energy_test = energy_test.apply(calculate_model, axis=1, result_type="expand")
energy_test_lr_rmsle = np.sqrt(energy_test["meter_reading_lr_q"].sum() / len_)
energy_test_mean_rmsle = np.sqrt(energy_test["meter_reading_mean_q"].sum() / len_)

print(f"Качество среднего: {energy_test_mean_rmsle}")
print(f"Качество линейной регрессии: {energy_test_lr_rmsle}")
