==> data-science-algo/tasks/9_model_hierarchy.ipynb <==
# Иерархия моделей

1. Посчитать модель линейной регрессии по первым 100 зданиям.
2. Найти точность, используя только дни недели и праздники, примемяя fit_intercept=False и логарифмируя целевой показатель.

Для вычисления отсутствующих или некорректных данных построить модели по всем зданиям одного типа в одном и во всех городах.

## Подключение библиотек

import pandas as pd
import numpy as np

from pandas.tseries.holiday import USFederalHolidayCalendar as calendar
from scipy.interpolate import interp1d
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression

from core.reduce_mem_usage import reduce_mem_usage

## Загрузка данных

buildings = pd.read_csv("../data/buildings.csv")
weather = pd.read_csv("../data/weather.csv")
energy = pd.read_csv("../data/train.0.zip")

## Отбор 100 зданий, объединение и оптимизация

energy = energy[energy["building_id"]<100]
energy = pd.merge(
    left=energy,
    right=buildings,
    how="left",
    left_on="building_id",
    right_on="building_id",
)
energy = energy.set_index(["timestamp", "site_id"])
weather = weather.set_index(["timestamp", "site_id"])
energy = pd.merge(
    left=energy,
    right=weather,
    how="left",
    left_index=True,
    right_index=True,
)
energy.reset_index(inplace=True)
energy = energy.drop(
    columns=[
        "meter",
        "year_built",
        "square_feet",
        "floor_count",
    ],
    axis=1
)

del buildings
del weather

energy = reduce_mem_usage(energy)
print(energy.info())

## Обогащение данных: час, дни недели, праздники

energy["hour"] = energy["timestamp"].dt.hour.astype("int8")
energy["weekday"] = energy["timestamp"].dt.weekday.astype("int8")

for weekday in range(0, 7):
    energy[f"is_wday {str(weekday)}"] = energy["weekday"].isin([weekday]).astype("int8")

energy["date"] = pd.to_datetime(energy["timestamp"].dt.date)
dates_range = pd.date_range(start="2015-12-31", end="2017-01-01")
us_holidays = calendar().holidays(start=dates_range.min(), end=dates_range.max())
energy["is_holiday"] = energy["date"].isin(us_holidays).astype("int8")

## Логарифмирование данных

energy["meter_reading_log"] = np.log(energy["meter_reading"] + 1)

## Разделение данных

energy_train, energy_test = train_test_split(
    energy[energy["meter_reading"] > 0],
    test_size=0.2
)
print(energy_train.info())

## Линейная регрессия:

1. по часам

hours = range(0, 24)
buildings = range(0, energy_train["building_id"].max() + 1)
lr_columns = [
    "meter_reading_log",
    "hour",
    "building_id",
    "is_holiday",
]

for wday in range(0, 7):
    lr_columns.append(f"is_wday {str(wday)}")

energy_train_lr = pd.DataFrame(energy_train, columns=lr_columns)
energy_lr = [[] for _ in range(len(buildings))]

for building in buildings:
    energy_lr[building] = [[] for _ in range(len(hours))]
    energy_train_b = energy_train_lr[energy_train_lr["building_id"] == building]
    for hour in hours:
        energy_lr[building].append([0 for _ in range(len(lr_columns) - 3)])
        energy_train_bh = pd.DataFrame(energy_train_b[energy_train_b["hour"] == hour])
        y = energy_train_bh["meter_reading_log"]
        if len(y) > 0:
            x = energy_train_bh.drop(labels=[
                    "meter_reading_log",
                    "hour",
                    "building_id",
                ],
                axis=1
            )
            model = LinearRegression(fit_intercept=False).fit(x, y)
            energy_lr[building][hour] = model.coef_
            energy_lr[building][hour] = np.append(energy_lr[building][hour], model.intercept_)

print(energy_lr[0])

2. по типам зданий

sites = range(0, energy["site_id"].max() + 1)
primary_uses = energy["primary_use"].unique()
lr_columns_use = [
    "meter_reading_log",
    "hour",
    "building_id",
    "is_holiday",
    "primary_use",
    "site_id",
]

for wday in range(0, 7):
    lr_columns_use.append(f"is_wday {str(wday)}")

energy_lr_use = {}
energy_lr_use_site = {}
energy_train_lr = pd.DataFrame(energy_train, columns=lr_columns_use)

for primary_use in primary_uses:
    energy_train_u = energy_train_lr[energy_train_lr["primary_use"] == primary_use]
    if len(energy_train_u) > 0:
        energy_lr_use_site[primary_use] = [[] for _ in range(len(sites))]
        for site in sites:
            energy_lr_use_site[primary_use][site] = [[] for _ in range(len(hours))]
            energy_train_us = energy_train_u[energy_train_u["site_id"] == site]
            if len(energy_train_us) > 0:
                for hour in hours:
                    energy_train_uth = energy_train_us[energy_train_us["hour"] == hour]
                    y = energy_train_uth["meter_reading_log"]
                    if len(y) > 0:
                        x = energy_train_uth.drop(
                            labels=[
                                "meter_reading_log",
                                "hour",
                                "building_id",
                                "site_id",
                                "primary_use",
                            ],
                            axis=1
                        )
                        model = LinearRegression(fit_intercept=False).fit(x, y)
                        energy_lr_use_site[primary_use][site][hour] = model.coef_
                        energy_lr_use_site[primary_use][site][hour] = np.append(energy_lr_use_site[primary_use][site][hour], model.intercept_)
        energy_lr_use[primary_use] = [[] for _ in range(len(hours))]
        for hour in hours:
            energy_train_th = energy_train_u[energy_train_u["hour"] == hour]
            y = energy_train_th["meter_reading_log"]
            if len(y) > 0:
                x = energy_train_th.drop(
                    labels=[
                        "meter_reading_log",
                        "hour",
                        "building_id",
                        "site_id",
                        "primary_use",
                    ],
                    axis=1
                )
                model = LinearRegression(fit_intercept=False).fit(x, y)
                energy_lr_use[primary_use][hour] = model.coef_
                energy_lr_use[primary_use][hour] = np.append(energy_lr_use[primary_use][hour], model.intercept_)

print(energy_lr_use_site)

## Расчет качества

def calculate_model(x):
    lr = -1
    model = energy_lr[x.building_id][x.hour]
    if len(model) == 0:
        model = energy_lr_use_site[x.primary_use][x.site_id][x.hour]
    if len(model) == 0:
        model = energy_lr_use[x.primary_use][x.hour]
    if len(model) > 0:
        lr = np.sum([x[col] * model[i] for i, col in enumerate(lr_columns[3:])])
        lr += model[len(lr_columns) - 3]
        lr = np.exp(lr)
    if lr < 0:
        lr = 0
    
    x["meter_reading_lr_q"] = (np.log(x.meter_reading + 1) - np.log(1 + lr))**2

    return x

energy_test = energy_test.apply(calculate_model, axis=1, result_type="expand")
energy_test_lr_rmsle = np.sqrt(energy_test["meter_reading_lr_q"].sum() / len(energy_test))
print(f"Качество линейной регрессии, 20 зданий: {energy_test_lr_rmsle}, (округленно): {round(energy_test_lr_rmsle, 1)}")
