==> data-science-algo/tasks/10_regression_optimization.ipynb <==
# Оптимизация регрессии

Несколько моделей линейной регрессии, что бы найти более оптимальную для первых 20 зданий

import pandas as pd
import numpy as np

from pandas.tseries.holiday import USFederalHolidayCalendar as calendar
from scipy.interpolate import interp1d
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression, Lasso, Ridge, ElasticNet, BayesianRidge

from core.reduce_mem_usage import reduce_mem_usage

## Загрузка данных

buildings = pd.read_csv("../data/buildings.csv")
weather = pd.read_csv("../data/weather.csv")
energy = pd.read_csv("../data/train.0.zip")

## Взять 20 зданий, объединить, оптимизировать

energy = energy[energy["building_id"] < 20]
energy = pd.merge(
    left=energy,
    right=buildings,
    how="left",
    left_on="building_id",
    right_on="building_id",
)
energy = energy.set_index(["timestamp", "site_id"])
weather =weather.set_index(["timestamp", "site_id"])
energy = pd.merge(
    left=energy,
    right=weather,
    how="left",
    left_index=True,
    right_index=True,
)
energy.reset_index(inplace=True)
energy = energy.drop(
    columns=[
        "meter",
        "site_id",
        "year_built",
        "square_feet",
        "floor_count",
    ],
    axis=1
)

del buildings
del weather

energy = reduce_mem_usage(energy)
print(energy.info())

## Обогащение данных: час, дни недели, праздники

energy["hour"] = energy["timestamp"].dt.hour.astype("int8")
energy["weekday"] = energy["timestamp"].dt.weekday.astype("int8")

for weekday in range(0, 7):
    energy[f"is_wday {str(weekday)}"] = energy["weekday"].isin([weekday]).astype("int8")

energy["date"] = pd.to_datetime(energy["timestamp"].dt.date)
dates_range = pd.date_range(start="2015-12-31", end="2017-01-01")
us_holidays = calendar().holidays(start=dates_range.min(), end=dates_range.max())
energy["is_holiday"] = energy["date"].isin(us_holidays).astype("int8")

## Логарифм

energy["meter_reading_log"] = np.log(energy["meter_reading"] + 1)

## Разделение данных

energy_train, energy_test = train_test_split(
    energy[energy["meter_reading"] > 0],
    test_size=0.2
)
print(energy_train.info())

energy_train.head()

## Линейная регрессия: по часам

from sklearn.metrics import r2_score

hours = range(0, 24)
buildings = range(0, energy_train["building_id"].max() + 1)
lr_columns = [
    "meter_reading_log",
    "hour",
    "building_id",
    "is_holiday",
]

for wday in range(0, 7):
    lr_columns.append(f"is_wday {str(wday)}")

lr_models = {
    "LinearRegression": LinearRegression,
    "Lasso-0.01": Lasso,
    "Lasso-0.1": Lasso,
    "Lasso-1.0": Lasso,
    "Ridge-0.01": Ridge,
    "Ridge-0.1": Ridge,
    "Ridge-1.0": Ridge,
    "ElasticNet-1-1": ElasticNet,
    "ElasticNet-0.1-1": ElasticNet,
    "ElasticNet-1-0.1": ElasticNet,
    "ElasticNet-0.1-0.1": ElasticNet,
    "BayesianRidge": BayesianRidge,
}

energy_train_lr = pd.DataFrame(energy_train, columns=lr_columns)

Линейная регрессия:
$$\LARGE z = Ax + By + C, \lvert z + z_0 \rvert^2 \to min$$
Лассо + LARS Лассо:
$$\LARGE \dfrac{1}{2n} \lvert z - z_0 \rvert^2 + a(\lvert A \rvert + \lvert B \rvert) \to min$$
Гребная регрессия:
$$\LARGE \lvert z - z_0 \rvert^2 + a(A^2 + B^2) \to min$$
ElasticNet: Лассо + Гребневая регрессия:
$$\LARGE \dfrac{1}{2n} \lvert z - z_0 \rvert^2 + ap \lvert A^2 + B^2\rvert + (a - p)(\lvert A \rvert + \lvert B \rvert)/2 \to min$$

lr_models_scores = {}

for i in lr_models:
    lr_model = lr_models[i]
    energy_lr_scores = [[] for _ in range(len(buildings))]

    for building in buildings:
        energy_lr_scores[building] = [0 for _ in range(len(hours))]
        energy_train_b = energy_train_lr[energy_train_lr["building_id"] == building]
        
        for hour in hours:
            energy_train_bh = energy_train_b[energy_train_b["hour"] == hour]
            y = energy_train_bh["meter_reading_log"]
            x = energy_train_bh.drop(
                labels=[
                    "meter_reading_log",
                    "hour",
                    "building_id",
                ],
                axis=1
            )
            if i in ["Ridge-0.1", "Lasso-0.1"]:
                model = lr_model(alpha=.1, fit_intercept=False).fit(x, y)
            elif i in ["Ridge-0.01", "Lasso-0.01"]:
                model = lr_model(alpha=.01, fit_intercept=False).fit(x, y)
            elif i == "ElasticNet-1-1":
                model = lr_model(alpha=1, l1_ratio=1, fit_intercept=False).fit(x, y)
            elif i == "ElasticNet-1-0.1":
                model = lr_model(alpha=1, l1_ratio=.1, fit_intercept=False).fit(x, y)
            elif i == "ElasticNet-0.1-1":
                model = lr_model(alpha=.1, l1_ratio=1, fit_intercept=False).fit(x, y)
            elif i == "ElasticNet-0.1-0.1":
                model = lr_model(alpha=.1, l1_ratio=.05, fit_intercept=False).fit(x, y)
            else:
                model = lr_model(fit_intercept=False).fit(x, y)
            
            energy_lr_scores[building][hour] = r2_score(y, model.predict(x))
        
    lr_models_scores[i] = np.mean(energy_lr_scores)

print(lr_models_scores)

## Проверим модели: LinearRegression, Lasso, BayesianRidge

energy_lr = [[] for _ in range(len(buildings))]
energy_lasso = [[] for _ in range(len(buildings))]
energy_br = [[] for _ in range(len(buildings))]

for building in buildings:
    energy_train_b = energy_train_lr[energy_train_lr["building_id"] == building]
    energy_lr[building] = [[] for _ in range(len(hours))]
    energy_lasso[building] = [[] for _ in range(len(hours))]
    energy_br[building] = [[] for _ in range(len(hours))]

    for hour in hours:
        energy_train_bh = pd.DataFrame(energy_train_b[energy_train_b["hour"] == hour])
        y = energy_train_bh["meter_reading_log"]
        if len(y) > 0:
            x = energy_train_bh.drop(
                labels=[
                    "meter_reading_log",
                    "hour",
                    "building_id",
                ],
                axis=1
            )
        model = LinearRegression(fit_intercept=False).fit(x, y)
        energy_lr[building][hour] = np.append([model.coef_], model.intercept_)
        
        model = Lasso(fit_intercept=False, alpha=.01).fit(x, y)
        energy_lasso[building][hour] = np.append([model.coef_], model.intercept_)
        
        model = BayesianRidge(fit_intercept=False).fit(x, y)
        energy_br[building][hour] = np.append([model.coef_], model.intercept_)

print(energy_lr[0][0])
print(energy_lasso[0][0])
print(energy_br[0][0])
