==> data-science-algo/tasks/11_data_export_and_import.ipynb <==
# Экспорт и импорт данных

1. Подготовить данные для построения модели.
2. Сохранить готовые данные в: CSV, HDF5

import pandas as pd
import numpy as np

from os import getcwd
from os.path import getsize
from pandas.tseries.holiday import USFederalHolidayCalendar as calendar
from scipy.interpolate import interp1d
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression

from core.reduce_mem_usage import reduce_mem_usage

## Загрузка данных

buildings = pd.read_csv("../data/buildings.csv")
weather = pd.read_csv("../data/weather.csv")
energy = pd.read_csv("../data/train.0.zip")

## Взять 20 зданий, объединить, оптимизировать

weather = weather[weather["site_id"] == 0]
energy = energy[energy["building_id"] < 20]
energy = pd.merge(
    left=energy,
    right=buildings,
    how="left",
    left_on="building_id",
    right_on="building_id",
)
del buildings

## Интерполяция значений

weather["precip_depth_1_hr"] = weather["precip_depth_1_hr"].apply(lambda x: x if x> 0 else 0)
interpolate_columns = [
    "air_temperature",
    "dew_temperature",
    "cloud_coverage",
    "wind_speed",
    "wind_direction",
    "precip_depth_1_hr",
    "sea_level_pressure",
]

for col in interpolate_columns:
    weather[col] = weather[col].interpolate(limit_direction="both", kind="cubic")

## Обогащение данных: погода

weather["air_temperature_diff_1"] = weather["air_temperature"].diff()
weather.at[0, "air_temperature_diff_1"] = weather.at[1, "air_temperature_diff_1"]
weather["air_temperature_diff_2"] = weather["air_temperature_diff_1"].diff()
weather.at[0, "air_temperature_diff_2"] = weather.at[1, "air_temperature_diff_2"]

## Объединение погодных данных

energy = energy.set_index(["timestamp", "site_id"])
weather = weather.set_index(["timestamp", "site_id"])
energy = pd.merge(
    left=energy,
    right=weather,
    how="left",
    left_index=True,
    right_index=True,
)
energy.reset_index(inplace=True)
energy = energy.drop(
    columns=[
        "meter",
        "site_id",
        "year_built",
        "square_feet",
        "floor_count",
    ], 
    axis=1
)
del weather
energy = reduce_mem_usage(energy)
print(energy.info())

## Обогащение данных: дата

energy["hour"] = energy["timestamp"].dt.hour.astype("int8")
energy["weekday"] = energy["timestamp"].dt.weekday.astype("int8")
energy["week"] = energy["timestamp"].dt.isocalendar().week.astype("int8")
energy["month"] = energy["timestamp"].dt.month.astype("int8")
energy["date"] = pd.to_datetime(energy["timestamp"].dt.date)

dates_range = pd.date_range(start="2015-12-31", end="2017-01-01")
us_holidays = calendar().holidays(start=dates_range.min(), end=dates_range.max())
energy["is_holiday"] = energy["date"].isin(us_holidays).astype("int8")

for weekday in range(0, 7):
    energy[f"is_wday {str(weekday)}"] = energy["weekday"].isin([weekday]).astype("int8")
for week in range(1, 54):
    energy[f"is_w {str(week)}"] = energy["week"].isin([week]).astype("int8")
for month in range(1, 13):
    energy[f"is_m {str(month)}"] = energy["month"].isin([month]).astype("int8")

## Логарифмирование данных

energy["meter_reading_log"] = np.log(energy["meter_reading"] + 1)

## Экспорт в CSV

print(energy.info())
energy.to_csv("../data/out/energy.0-20.csv.gz", index=False)

energy = pd.read_csv("../data/out/energy.0-20.csv.gz")
print(energy.info())

## Экспорт в HDF5

energy = reduce_mem_usage(energy)
energy.to_hdf(
    "../data/out/energy.0-20.h5",
    key="energy",
    format="table",
    complib="zlib",
    complevel=9,
    mode="w"
)


print(f"CSV: {getsize('../data/out/energy.0-20.csv.gz')}")
print(f"HDF5: {getsize('../data/out/energy.0-20.h5')}")

energy = pd.read_hdf("../data/out/energy.0-20.h5", "energy")
print(energy.info())

## Разделение данных

energy_train, energy_test = train_test_split(energy[energy["meter_reading"] > 0], test_size=0.2)
print(energy_train.head())

pd.set_option("io.hdf.default_format", "table")
store = pd.HDFStore("../data/out/energy.0-20.split.h5",
                    complevel=9,
                    complib="zlib",
                    mode="w",
                    )
store["energy_train"] = energy_train
store["energy_test"] = energy_test
store.put("metadata",
          pd.Series(["Набор обогащенных тестовых данных по 20 зданиям"]))
store.close()

print(f"HDF5: {getsize('../data/out/energy.0-20.split.h5')}")

## Чтение из HDF5

store = pd.HDFStore("../data/out/energy.0-20.split.h5")
energy_test = store.get("energy_test")[:]
energy_train = store.get("energy_train")[:]
metadata = store.get("metadata")[:]
store.close()
print(metadata[0])
print(energy_train.head())
