==> data-science-algo/tasks/13_calculation_of_results.ipynb <==
# Расчет результатов

1. Посчитать модели линейной регрессии для 20 зданий по оптимальному набору параметров: метеорологические данные, дни недели, недели года, месяцы и праздники по всему набору данных.
2. Загрузить данные решения, посчитать значение энергопотребления для требуемых дат для тех зданий, которые посчитаны в модели, и выгрузить результат в файл.

## Подключение библиотек

import pandas as pd
import numpy as np

from pandas.tseries.holiday import USFederalHolidayCalendar as calendar
from scipy.interpolate import interp1d
from sklearn.linear_model import LinearRegression

from core.reduce_mem_usage import reduce_mem_usage

## Загрузка данных 20 зданий из HDF5

energy = pd.read_hdf("../data/out/energy.0-20.h5", "energy")
print(energy.info())

## Загрузка данных и оптимизация памяти

buidings = pd.read_csv("../data/buildings.csv", usecols=["site_id", "building_id"])

weather = pd.read_csv("../data/weather_test.csv.gz")
weather = weather[weather["site_id"] == 0]
weather = weather.drop(columns=["wind_direction"], axis=1)

results = pd.read_csv("../data/test.parquet.gzip")
results = results[(results["building_id"] < 20) & (results["meter"] == 0)]
results = pd.merge(
    left=results,
    right=buidings,
    how="left",
    left_on="building_id",
    right_on="building_id",
)

del buidings

results = results.drop(columns=["meter"], axis=1)
print(results.info())

## Интерполяция значений и обогащение погодных данных для одного года

interpolate_columns = [
    "air_temperature",
    "dew_temperature",
    "cloud_coverage",
    "wind_speed",
    "sea_level_pressure",
]

for col in interpolate_columns:
    weather[col] = weather[col].interpolate(limit_direction="both", kind="cubic")

weather["air_temperature_diff_1"] = weather["air_temperature"].diff()
weather.at[0, "air_temperature_diff_1"] = weather.at[1, "air_temperature_diff_1"]
weather["air_temperature_diff_2"] = weather["air_temperature_diff_1"].diff()
weather.at[0, "air_temperature_diff_2"] = weather.at[1, "air_temperature_diff_2"]

## Объединение данных по погоде

results = results.set_index(["timestamp", "site_id"])
weather = weather.set_index(["timestamp", "site_id"])

results = pd.merge(
    left=results,
    right=weather,
    how="left",
    left_index=True,
    right_index=True,
)

results.reset_index(inplace=True)
results = results.drop(columns=["site_id"], axis=1)

del weather

results = reduce_mem_usage(results)
print(results.info())

## Обогащение данных по дате

results["hour"] = results["timestamp"].dt.hour.astype("int8")
results["weekday"] =results["timestamp"].dt.weekday.astype("int8")
results["week"] =results["timestamp"].dt.isocalendar().week.astype("int8")
results["month"] =results["timestamp"].dt.month.astype("int8")
results["date"] = pd.to_datetime(results["timestamp"].dt.date)

dates_range = pd.date_range(start="2016-12-31", end="2018-06-01")
us_holidays = calendar().holidays(start=dates_range.min(), end=dates_range.max())
results["is_holiday"] = results["date"].isin(us_holidays).astype("int8")

for weekday in range(0, 7):
    results[f"is_wday {str(weekday)}"] = results["weekday"].isin([weekday]).astype("int8")
for week in range(1, 54):
    results[f"is_w {str(week)}"] = results["week"].isin([week]).astype("int8")
for month in range(1, 13):
    results[f"is_m {str(month)}"] = results["month"].isin([month]).astype("int8")

## Линейная регрессия

hours = range(0, 24)
buildings = range(0, energy["building_id"].max() + 1)
lr_columns = [
    "meter_reading_log",
    "hour",
    "building_id",
    "air_temperature",
    "dew_temperature",
    "sea_level_pressure",
    "wind_speed",
    "cloud_coverage",
    "air_temperature_diff_1",
    "air_temperature_diff_2",
    "is_holiday",
]

for wday in range(0, 7):
    lr_columns.append(f"is_wday {str(wday)}")
for week in range(1, 54):
    lr_columns.append(f"is_w {str(week)}")
for month in range(1, 13):
    lr_columns.append(f"is_m {str(month)}")

energy_train_lr = pd.DataFrame(energy, columns=lr_columns)
energy_lr = [[] for _ in range(len(buildings))]

for building in buildings:
    energy_lr[building] = [[] for _ in range(len(hours))]
    energy_train_b = energy_train_lr[energy_train_lr["building_id"] == building]

    for hour in hours:
        energy_train_bh = energy_train_b[energy_train_b["hour"] == hour]
        y = energy_train_bh["meter_reading_log"]
        x = energy_train_bh.drop(
            labels=["meter_reading_log",
                    "hour",
                    "building_id"],
            axis=1
        )
        model = LinearRegression(fit_intercept=False).fit(x, y)
        energy_lr[building][hour] = model.coef_
        energy_lr[building][hour] = np.append(
            energy_lr[building][hour],
            model.intercept_
        )

print(energy_lr[0])

## Расчет финальных показателей, только энергопотребление 20 первых зданий

def calculate_model(x):
    lr = -1
    model = energy_lr[x.building_id][x.hour]
    if len(model) > 0:
        lr = np.sum([x[col] * model[i] for i, col in enumerate(lr_columns[3:])])
        lr +=model[len(lr_columns) - 3]
        lr = np.exp(lr)
    if lr < 0 or lr != lr or lr*lr == lr:
        lr =0
    x["meter_reading"] = lr
    if x["row_id"] % 1000 == 0:
        print(f"Готово {x['row_id']}")
    return x

results = results.apply(calculate_model, axis=1, result_type="expand")

## Усечение данных до требуемого формата

results_ready = pd.DataFrame(results, columns=["row_id", "meter_reading"])

## Загрузка всех данных для заполнения их нулями

results = pd.read_csv("../data/test.parquet.gzip", usecols=["row_id"])
results = pd.merge(
    left=results,
    right=results_ready,
    how="left",
    left_on="row_id",
    right_on="row_id",
)
results.fillna(value=0, inplace=True)
print(results.info())

## Выгрузка результатов в CSV файл

results.to_csv("../data/out/submission.csv.gz", index=False, compression="gzip")

## Освободить память

del energy
del results
del results_ready
