==> data-science-algo/tasks/15_f1_and_kappa_score_classification.ipynb <==
# F1 и каппа оценка классификации

1. Разделить выборку на обуч./пров. в соотн. 80/20
2. Для обучения вычислить среднее значение для веса, роста, индекса массы тела и возраста - для каждого из принятых решений. Предсказать оценку скоринга по близости данных средним значениям.
3. Проверить качество предсказаний через F1-метрику и матрицу неточностей.

## Подключение библиотек

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.metrics import f1_score, confusion_matrix, cohen_kappa_score

## Загрузка данных

data = pd.read_csv("../data/prudential/train.csv.gz")
print(data.info())

## Разделить данные на обучающие и проверочные, 80/20

data_train, data_test = train_test_split(data, test_size=0.2)
print(data_train.head())

## Вычислить медиану для каждой оценки

columns = [
    "Wt",
    "Ht",
    "Ins_Age",
    "BMI",
]

responses = np.arange(1, data["Response"].max() + 1)
clusters = [{} for _ in range(len(responses) + 1)]

for r in responses:
    for c in columns:
        clusters[r][c] = data[data["Response"] == r][c].median()

print(clusters)

## Выполнить предсказание оценки скоринга на основе медианы

Используем евклидово расстояние:
$$ D=\sqrt{\sum(a_i - C_i)^2} $$


def calc_model(x, clusters, columns, data):
    D_min = 10000000
    target = 0
    for i, cluster in enumerate(clusters):
        if cluster:
            D = 0
            for c in columns:
                D += (cluster[c] - x[c])**2
            D = np.sqrt(D)
            if D < D_min:
                target = i
                D_min = D
    x["target"] = target
    x["random"] = int(np.random.uniform(1, 8.01, 1)[0])
    x["sample"] = data.sample(1)["Response"].values[0]
    x["all8"] = 8
    return x

data_test = data_test.apply(calc_model, axis=1, result_type="expand", args=(clusters, columns, data))
print(data_test.head(20))

## Оценка качества модели: F1

print(f"Случайный выбор: {f1_score(data_test['random'], data_test['Response'], average='weighted')}")
print(f"Выбор по частоте: {f1_score(data_test['sample'], data_test['Response'], average='weighted')}")
print(f"Кластеризация: {f1_score(data_test['target'], data_test['Response'], average='weighted')}")
print(f"Самый популярный: {f1_score(data_test['all8'], data_test['Response'], average='weighted')}")

## Матрица неточностей

print(confusion_matrix(data_test["target"], data_test["Response"]))

## Квадратный коэффициент Каппа Коэна

print(cohen_kappa_score(data_test["target"], data_test["Response"], weights="quadratic"))
print(cohen_kappa_score(data_test["all8"], data_test["Response"], weights="quadratic"))
