==> data-science-algo/tasks/16_k-nearest_neighbors.ipynb <==
# Метод ближайших соседей

1. Разделить выборку 80/20
2. Применить метод ближайших соседей для классификации скоринга.
3. Проверить качество предсказания через каппа-метрику и матрицу неточностей.

## Подключить библиотеки

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.metrics import cohen_kappa_score, confusion_matrix
from sklearn.neighbors import KNeighborsClassifier

## Загрузка данных

data = pd.read_csv("../data/prudential/train.csv.gz")
print(data.info())

## Разделить данные

data_train, data_test = train_test_split(data, test_size=0.2)
print(data_train.head())

## Расчет модели kNN

Вычислить не центры(кластеры) исходных групп, а расстояния до всех значений. Выбираем то значение, кот. превалирует у k ближайших соседей.

Для оценки качества возьмем **k** равным **10**, **100**, **1000**, **10000**

columns = [
    "Wt",
    "Ht",
    "Ins_Age",
    "BMI",
]
max_nn = data_train.groupby("Response").count()["Id"].min()
knn_10 = KNeighborsClassifier(n_neighbors=10)
knn_100 = KNeighborsClassifier(n_neighbors=100)
knn_1000 = KNeighborsClassifier(n_neighbors=1000)
knn_10000 = KNeighborsClassifier(n_neighbors=10000)
knn_max = KNeighborsClassifier(n_neighbors=max_nn)

y = data_train["Response"]
x = pd.DataFrame(data_train, columns=columns)
knn_10.fit(x, y)
knn_100.fit(x, y)
knn_1000.fit(x, y)
knn_10000.fit(x, y)
knn_max.fit(x, y)

## Предсказание данных

x_test = pd.DataFrame(data_test, columns=columns)
data_test["target_10"] = knn_10.predict(x_test)
data_test["target_100"] = knn_100.predict(x_test)
data_test["target_1000"] = knn_1000.predict(x_test)
data_test["target_10000"] = knn_10000.predict(x_test)
data_test["target_max"] = knn_max.predict(x_test)
print(data_test.head(20))

## Оценка модели

print(f'kNN, 10: {cohen_kappa_score(data_test["target_10"], data_test["Response"], weights="quadratic")}')
print(f'kNN, 100: {cohen_kappa_score(data_test["target_100"], data_test["Response"], weights="quadratic")}')
print(f'kNN, 1000: {cohen_kappa_score(data_test["target_1000"], data_test["Response"], weights="quadratic")}')
print(f'kNN, 10000: {cohen_kappa_score(data_test["target_10000"], data_test["Response"], weights="quadratic")}')
print(f'kNN, max: {cohen_kappa_score(data_test["target_max"], data_test["Response"], weights="quadratic")}')

## Матрицы неточностей

print(confusion_matrix(data_test["target_10"], data_test["Response"]))

print(confusion_matrix(data_test["target_10000"], data_test["Response"]))
