==> data-science-algo/tasks/17_naiven_bayes.ipynb <==
# Наивный Байес

1. Применить наивный Байес для классификации скоринга. Использовать все столбцы.
2. Проверить качество через каппа-метрику и матрицу неточностей.

## Подключение библиотек

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.metrics import cohen_kappa_score, confusion_matrix
from sklearn.naive_bayes import GaussianNB
from sklearn import preprocessing

from core.reduce_mem_usage import reduce_mem_usage

## Загрузка данных

data = pd.read_csv("../data/prudential/train.csv.gz")
print(data.info())

## Категоризация данных

data["Product_Info_2_1"] = data["Product_Info_2"].str.slice(0, 1)
data["Product_Info_2_2"] = pd.to_numeric(data["Product_Info_2"].str.slice(1, 2))
data.drop(labels=["Product_Info_2"], axis=1, inplace=True)
print(data.info())

## Оптимизация памяти

data = reduce_mem_usage(data)
print(data.info())

## Предобработка

list_df = [data]
for un in data["Product_Info_2_1"].unique():
    s = pd.DataFrame()
    s["Product_Info_2_1_" + un] = data["Product_Info_2_1"].isin([un]).astype("int8")
    list_df.append(s)
data = pd.concat(list_df, axis=1)
data.drop(labels=["Product_Info_2_1"], axis=1, inplace=True)

## Заполнить отсутствующие значения
-1 увеличивает расстояние при расчёте ближайших соседей

data.fillna(value=-1, inplace=True)

## Столбцы для модели

columns_groups = [
    "Insurance_History",
    "InsuredInfo",
    "Medical_Keyword",
    "Family_Hist",
    "Medical_History",
    "Product_Info",
]
columns = [
    "Wt",
    "Ht",
    "Ins_Age",
    "BMI",
]

for cg in columns_groups:
    columns.extend(data.columns[data.columns.str.startswith(cg)])
print(columns)

## Z-нормализация данных

scaler = preprocessing.StandardScaler()
scaler.fit(pd.DataFrame(data, columns=columns))

## Разделение данных

data_train, data_test = train_test_split(data, test_size=0.2)
print(data_train.head())

## Расчет модели наивного Байеса

$$ P(A|B) = \dfrac{P(B|A)*P(A)}{P(B)} $$

y = data_train["Response"]
x = scaler.transform(pd.DataFrame(data_train, columns=columns))

bayes = GaussianNB()
bayes.fit(x, y)

## Предсказание данных

data_test = pd.DataFrame(data_test)
x_test = scaler.transform(pd.DataFrame(data_test, columns=columns))
data_test["target"] = bayes.predict(x_test)
print(data_test.head())

## Оценка модели

print(f'Байес: {cohen_kappa_score(data_test["target"], data_test["Response"], weights="quadratic")}')

## Матрица неточностей

print(confusion_matrix(data_test["target"], data_test["Response"]))
