==> data-science-algo/tasks/19_hierarchy_of_logical_regression..ipynb <==
# Иерархия логистической регрессии

Построить 4 модели логистической регрессии:
1. для 8, 6 и остальных классов
2. для 2, 5 и остальных классов
3. для 1, 7 и остальных классов
4. для 3 и 4 -по убыванию частоты значения

Использовать перекрестную проверку при принятии решения об оптимальном наборе столбцов

Проверить предсказание через каппа-метрику.

## Подключить библиотеки

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score
from sklearn.metrics import cohen_kappa_score, confusion_matrix, make_scorer
from sklearn.linear_model import LogisticRegression
from sklearn import preprocessing

from core.reduce_mem_usage import reduce_mem_usage

## Загрузка данных

data = pd.read_csv("../data/prudential/train.csv.gz")
print(data.info())

## Предобработка данных

data["Product_Info_2_1"] = data["Product_Info_2"].str.slice(0, 1)
data["Product_Info_2_2"] = pd.to_numeric(data["Product_Info_2"].str.slice(1, 2))
data.drop(labels=["Product_Info_2"], axis=1, inplace=True)

list_df = [data]
for un in data["Product_Info_2_1"].unique():
    s = pd.DataFrame()
    s["Product_Info_2_1_" + un] = data["Product_Info_2_1"].isin([un]).astype("int8")
    list_df.append(s)
data = pd.concat(list_df, axis=1)
data.drop(labels=["Product_Info_2_1"], axis=1, inplace=True)

data.fillna(value=-1, inplace=True)
print(data.head())

## Оптимизировать память

data = reduce_mem_usage(data)
print(data.info())

## Набор столбцов для расчета

columns_groups = [
    "Insurance_History",
    "InsuredInfo",
    "Medical_Keyword",
    "Family_Hist",
    "Medical_History",
    "Product_Info",
]
columns = [
    "Wt",
    "Ht",
    "Ins_Age",
    "BMI",
]

for cg in columns_groups:
    columns.extend(data.columns[data.columns.str.startswith(cg)])
print(columns)

## Предобработка данных

scaler = preprocessing.StandardScaler()
data_transformed = pd.DataFrame(scaler.fit_transform(pd.DataFrame(data, columns=columns)))
columns_transformed = data_transformed.columns
data_transformed["Response"] = data["Response"]

## Разделение данных

data_train, data_test = train_test_split(data_transformed, test_size=0.2)
print(data_train.head())

## Логистическая регрессия

В обучающих данных пометим все классы, кроме 6 и 8, как 0 - и проведем обучение.
Затем в оставшихся данных (в кот класс не равен 6 или 8) заменим все классы, кроме 7 и 1, на 0 - и снова проведем обучение и т.д.

### Оптимальный набор столбцов
Для каждого уровня иерархии это будет свой набор столбцов в исходных данных.

### Перекрестная проверка
Разбиваем обущающую выборку на k частей. На каждой части обучаем модель. Затем проверяем 1-ю, 2-ю, 3-ю, 4-ю на 5-ой; 1-ю, 2-ю, 3-ю, 5-ю на 4-ой и т.д.

def regression_model(df, columns):
    x = pd.DataFrame(df, columns=columns)
    model = LogisticRegression(max_iter=1000)
    model.fit(x, df["Response"])
    return model

def logistic_regression(df_train, columns):
    model = regression_model(df_train, columns)
    logr_grid = GridSearchCV(model, {}, cv=5, n_jobs=2, scoring=make_scorer(cohen_kappa_score))
    x = pd.DataFrame(df_train, columns=columns)
    logr_grid.fit(x, df_train["Response"])
    return logr_grid.best_score_

def find_opt_columns(data_train):
    kappa_score_opt = 0
    columns_opt = []
    for col in columns_transformed:
        kappa_score = logistic_regression(data_train, [col])
        if kappa_score > kappa_score_opt:
            columns_opt = [col]
            kappa_score_opt = kappa_score
    for col in columns_transformed:
        if col not in columns_opt:
            columns_opt.append(col)
            kappa_score = logistic_regression(data_train, columns_opt)
            if kappa_score < kappa_score_opt:
                columns_opt.pop()
            else:
                kappa_score_opt = kappa_score
    return columns_opt, kappa_score_opt

Будем последовательно урезать набор данных при расчете более глубоких моделей: после получения разделения на 8 отсечем все данные со значение 8 и т.д.
После каждого расчета модели будем вычислять значения в проверочной выборке. Проверочную выборку нулями заполнять не будем.
Набор разделений 6/8, 2/5, 1/7, 3/4 дает наибольшую точность

responses = [
    [6, 8],
    [2, 5],
    [1, 7],
    [3, 4],
]

logr_models = [{} for _ in range(len(responses))]
data_train_current = data_train.copy()

for i, response in enumerate(responses):
    m_train = data_train_current.copy()
    if response != [3, 4]:
        m_train["Response"] = m_train["Response"].apply(lambda x: 0 if x not in response else x)
    columns_opt, kappa_score_opt = find_opt_columns(m_train)
    logr_models[i] = {
        "model": regression_model(m_train, columns_opt),
        "columns": columns_opt,
    }
    if response != [3, 4]:
        data_train_current = data_train_current[~data_train_current["Response"].isin(response)]
    
    print(f"{i} - {kappa_score_opt} - {columns_opt}", end="/n/n")

## Предсказание данных и оценка модели
Последовательно считаем предсказания для каждой классификации. После этого объединяем предсказание по иерархии.

def logr_hierarchy(x):
    for response in range(len(responses)):
        if x["target" + str(response)] > 0:
            x["target"] = x["target" + str(response)]
            break;
    return x

for response in range(len(responses)):
    model = logr_models[response]["model"]
    columns_opt = logr_models[response]["columns"]
    x = pd.DataFrame(data_test, columns=columns_opt)
    data_test["target" + str(response)] = model.predict(x)

data_test = data_test.apply(logr_hierarchy, axis=1, result_type="expand")
print(data_test.head())

print(f'{cohen_kappa_score(data_test["target"], data_test["Response"], weights="quadratic"), 3}')

## Матрица неточностей

print(confusion_matrix(data_test["target"], data_test["Response"]))
