==> data-science-algo/tasks/22_random_forest.ipynb <==
# Случайный лес

1. Построить параллельный ансамбль (бэггинг) решающих деревьев используя случайный лес.
2. Получить предсказания и проверить качество через каппа-метрику.

## Подключить библиотеки

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score
from sklearn.metrics import cohen_kappa_score, confusion_matrix, make_scorer
from sklearn.ensemble import RandomForestClassifier
from sklearn import preprocessing

from core.reduce_mem_usage import reduce_mem_usage

## Загрузка данных

data = pd.read_csv("../data/prudential/train.csv.gz")
print(data.info())

## Предобработка данных

data["Product_Info_2_1"] = data["Product_Info_2"].str.slice(0, 1)
data["Product_Info_2_2"] = pd.to_numeric(data["Product_Info_2"].str.slice(1, 2))
data.drop(labels=["Product_Info_2"], axis=1, inplace=True)

list_df = [data]
for un in data["Product_Info_2_1"].unique():
    s = pd.DataFrame()
    s["Product_Info_2_1_" + un] = data["Product_Info_2_1"].isin([un]).astype("int8")
    list_df.append(s)
data = pd.concat(list_df, axis=1)
data.drop(labels=["Product_Info_2_1"], axis=1, inplace=True)

data.fillna(value=-1, inplace=True)
print(data.head())

## Набор столбцов для расчета

columns_groups = [
    "Insurance_History",
    "InsuredInfo",
    "Medical_Keyword",
    "Family_Hist",
    "Medical_History",
    "Product_Info",
]
columns = [
    "Wt",
    "Ht",
    "Ins_Age",
    "BMI",
]

for cg in columns_groups:
    columns.extend(data.columns[data.columns.str.startswith(cg)])
print(columns)

## Нормализация данных

scaler = preprocessing.StandardScaler()
data_transformed = pd.DataFrame(scaler.fit_transform(pd.DataFrame(data, columns=columns)))
columns_transformed = data_transformed.columns
data_transformed["Response"] = data["Response"]

data_transformed = reduce_mem_usage(data_transformed)
print(data_transformed.info())

## Разделение данных

data_train, data_test = train_test_split(data_transformed, test_size=0.2)
print(data_train.head())

## Перекрестная проверка случайного леса

Каждое дерево (по умолчанию их 100) строиться на своей части выборки со своим набором параметров (max_features). Решение принимается путем голосования деревьев.

$$ estimators = \dfrac{N}{(20-100) * fold * class} $$

где:
- N - размер выборки
- fold - число разбиений
- class - количество классов в предсказании

x = pd.DataFrame(data_train, columns=columns_transformed)
model = RandomForestClassifier(
    random_state=7,
    n_estimators=77,
    max_depth=17,
    max_features=27,
    min_samples_leaf=30,
)

tree_params = {
    "max_depth": range(15, 17),
    "max_features": range(26, 28),
    "n_estimators": range(75, 77),
    "min_samples_leaf": range(19, 21),
} 

tree_grid = GridSearchCV(
    model,
    tree_params,
    cv=5,
    n_jobs=2,
    verbose=True, scoring=make_scorer(cohen_kappa_score)
)

tree_grid.fit(x, data_train["Response"])

## Оптимальные параметры

print(tree_grid.best_params_)

## Итоговая модель

model = RandomForestClassifier(
    random_state=17,
    min_samples_leaf=tree_grid.best_params_["min_samples_leaf"],
    max_features=tree_grid.best_params_["max_features"],
    max_depth=tree_grid.best_params_["max_depth"],
    n_estimators=tree_grid.best_params_["n_estimators"],
)

model.fit(x, data_train["Response"])

## Предсказание данных

x_test = pd.DataFrame(data_test, columns=columns_transformed)
data_test["target"] = model.predict(x_test)

## Оценка модели

print(f'Случайный лес: {round(cohen_kappa_score(data_test["target"], data_test["Response"], weights="quadratic"), 3)}')

## Матрица неточностей

print(f'Случайный лес:\n\n {confusion_matrix(data_test["target"], data_test["Response"])}')
