==> data-science-algo/tasks/26_image_preprocessing.ipynb <==
# Предобработка изображений

## Подключение библиотек

import pandas as pd
import numpy as np
from PIL import Image

## Загрузка данных

train = pd.read_csv("../data/clouds/train.csv.gz")
print(train.head())

## Очистка данных

train["Image"] = train["Image_Label"].str.split("_").str[0]
train["Label"] = train["Image_Label"].str.split("_").str[1]
train.drop(labels=["Image_Label"], axis=1, inplace=True)
print(train.head())

data = pd.DataFrame({"Image": train["Image"].unique()})
for label in train["Label"].unique():
    data[label] = pd.Series(train[train["Label"] == label]["EncodedPixels"].values)
print(data.head())

## Обработка изображений

0011165.jpg имеет размер 525 * 350 = 183750 пикселей

imgdata = np.array([np.zeros(183750, dtype="uint8") for _ in range(len(data))])

for i, img in enumerate(data["Image"].unique()):
    imgdata[i] = np.array(Image.open("../data/clouds/train_images_small/" + img).convert("L"),
                          dtype="uint8").reshape(1, -1)[0]
imgdata = pd.DataFrame(imgdata)
print(imgdata.head())

for column in data.columns:
    imgdata[column] = data[column]
del data
print(imgdata.head)

## Сохранить данные в HDF5

imgdata.to_hdf(
    "../data/out/clouds.data.h5",
    "data",
    complib="zlib",
    complevel=9,
    mode="w",
)
