==> data-science-algo/tasks/27_SVM_similarity_measure.ipynb <==
# Опорные векторы и коэфициент сходства

Для работы нужен файл clouds.data.h5. Если нет в папке - запустить 26_image_preprocessing.ipynb

1. Построить модель опорных векторов для типа облака Fish
2. Сделать оценку качества по F1 и коэфициенту сходства

## Подключение библиотек

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.linear_model import SGDClassifier
from sklearn.metrics import f1_score

## Загрузка данных

clouds = pd.read_hdf("../data/out/clouds.data.h5")
print(clouds.head())

## Оставить данные только по Fish

clouds.drop(
    labels=[
        "Image",
        "Flower",
        "Gravel",
        "Sugar",
    ],
    axis=1,
    inplace=True,
)

## Разделить данные

clouds_train, clouds_test = train_test_split(clouds, test_size=0.2)
del clouds
print(clouds_train.head())

## SVM

Считать последовательно коэффициенты для пакетов по 100 изображений

y = clouds_train["Fish"].notnull().astype("int8")
x = pd.DataFrame(clouds_train).drop(labels=["Fish"], axis=1)
model = SGDClassifier(loss="log", warm_start=True)

for i in range(len(clouds_train) // 100):
    model.partial_fit(x[i*100:i*100+100], y[i*100:i*100+100], classes=[0,1])

del x
del y

## Средняя область облаков

В качестве локализации облаков на изображении возьмем среднюю область по обучающей выборке

image_x = 2100
image_x_4 = image_x // 4
image_y =1400
image_y_4 = image_y // 4

def locate_rectangle(a):
    vals = [int(i) for i in a[0].split(" ")]
    x = vals[0] // image_y
    y = vals[0] % image_y
    width = (image_x + (vals[-2] + vals[-1]) // image_y - x) % image_x
    height = (vals[-2] + vals[-1]) % image_y - y
    return [x, y, width, height]

areas = pd.DataFrame(clouds_train["Fish"].copy().dropna(axis=0))
areas = areas.apply(locate_rectangle, axis=1, result_type="expand")
coords = np.array(areas.mean() // 4)
print(coords)

sgd_mask = np.zeros(image_x_4 * image_y_4,
                    dtype="uint8").reshape(image_x_4, image_y_4)
for x in range(image_x_4):
    for y in range(image_y_4):
        if (x >= coords[0] and x <= (coords[0] + coords[3]) and
                y >= coords[1] and y <= (coords[1] + coords[3])):
            sgd_mask[x][y] = 1
sgd_mask = sgd_mask.reshape(image_x_4 * image_y_4)
print(sgd_mask.sum())

## Предсказать значения

result = pd.DataFrame(
    {
        "EncodedPixels": clouds_test["Fish"],
        "Is_Fish": clouds_test["Fish"].notnull().astype("int8")
    }
)
result["target"] = model.predict(pd.DataFrame(clouds_test).drop(labels=["Fish"], axis=1))
print(result.head(10))

## Оценка предсказания F1

print(f'Опорные векторны: {round(f1_score(result["Is_Fish"], result["target"]), 3)}')
print(f'Все Fish: {round(f1_score(result["Is_Fish"], np.ones(len(result))), 3)}')

## Оценка по Дайсу

image_x = 525
image_y = 350


def mask_rate(a, x, y):
    b = a // 1400
    return np.round(x*(b*x // 2100) + y*(a%1400) // 1400).astype("uint32")

def calc_mask(px, x=525, y=350):
    p = np.array([int(n) for n in px.split(" ")]).reshape(-1, 2)
    mask = np.zeros(y*x, dtype="uint8")
    for i, m in p:
        mask[mask_rate(i, x, y)-1:mask_rate(m + i, x, y)] = 1
    return mask.reshape(y, x).transpose()

def calc_dice(x):
    dice = 0
    px = x["EncodedPixels"]
    if px != px and x["target"] == 0:
        dice = 1
    elif px == px and x["target"] == 1:
        mask = calc_mask(px).flatten()
        target = np.ones(image_x*image_y, dtype="uint8")
        dice = 2*np.sum(target[mask==1]) / (np.sum(target) + np.sum(mask))
    
    return dice

dice = result.apply(calc_dice, axis=1, result_type="expand")
print(f'Опорные векторы, Fish {round(dice.mean(), 3)}')
print(f'Опорные векторы, Fish {round(len(result[result["Is_Fish"] == 0]) / len(result), 3)}')
